requirements.txt
pydriller
pandas
numpy
matplotlib
seaborn
pyyaml
tqdm